Add transform for setting dependencies and fetches from upstream tasks
Up until now, downstream jobs in the pipeline have been 1:1 with their upstreams. Beginning with `merge-corpus`, this changes, and we now have a task that will consume a number of upstream tasks, and combine their outputs into one. Added to this is the fact we don't necessarily want _all_ of the upstream tasks for a given locale pair, we only want ones that:
* Use the best possible cleaning technique (differs by locale pair)
* Are for a dataset that is being requested in the training run (which comes in via parameters, usually through an Action task).

Because of this, we need to handle dependencies and fetches in this new transform.