Rework training action to take input more closely inline with existing config format.
Reference: https://github.com/mozilla/firefox-translations-training/blob/main/configs/config.test.yml

I've been meaning to do this for awhile, and I had to change this anyways due to merge corpus tasks not having a single `dataset` that we can generate a name from. This ends being a fairly significant rework to do a few things:
* Use the same format for `datasets` as the existing pipeline configs
* Use attributes to filter tasks in `train-target-tasks`, rather than the more brittle method of names.

The former necessitated amending the pipeline steps with a `dataset-category` attribute -- as there are multiple lists of datasets that the pipeline configures. Each part of the pipeline uses a different list.

There's still more to do to get this schema inline with the config, but this is a significant step forward, and unblocks the current work.