Add support for trg: en locale pairs
The only thing stopping this from working as far as the taskgraph is concerned was that we needed a bicleaner alias for the reverse locale pair.

I also ended up adjusting the prefix inside the archive. These packs come with a `$src-$trg` prefix, and it's much easier to script around them if we give them a consistent one.