Add merge corpus pipeline step to Taskcluster (#121)
* Enhancements to the `from_datasets` transform.

Two things:
1) Set a bunch of attributes to make things easier in downstream tasks, transforms, and target tasks methods.
2) Add a second transform sequence that collapses many upsrteams into a single job per locale pair. This is useful whenever we have a task that, eg: consumes the output of a number of upstream tasks that are per-dataset.

* Add support for zst in merge corpus pipeline script

This script got a slightly heavier tweak than the others to allow it to accept globs of files to process, rather than prefixes (it's much easier to provide this from Taskcluster).

* Add transform for setting dependencies and fetches from upstream tasks

Up until now, downstream jobs in the pipeline have been 1:1 with their upstreams. Beginning with `merge-corpus`, this changes, and we now have a task that will consume a number of upstream tasks, and combine their outputs into one. Added to this is the fact we don't necessarily want _all_ of the upstream tasks for a given locale pair, we only want ones that:
* Use the best possible cleaning technique (differs by locale pair)
* Are for a dataset that is being requested in the training run (which comes in via parameters, usually through an Action task).

Because of this, we need to handle dependencies and fetches in this new transform.

* Rework training action to take input more closely inline with existing config format.

Reference: https://github.com/mozilla/firefox-translations-training/blob/main/configs/config.test.yml

I've been meaning to do this for awhile, and I had to change this anyways due to merge corpus tasks not having a single `dataset` that we can generate a name from. This ends being a fairly significant rework to do a few things:
* Use the same format for `datasets` as the existing pipeline configs
* Use attributes to filter tasks in `train-target-tasks`, rather than the more brittle method of names.

The former necessitated amending the pipeline steps with a `dataset-category` attribute -- as there are multiple lists of datasets that the pipeline configures. Each part of the pipeline uses a different list.

There's still more to do to get this schema inline with the config, but this is a significant step forward, and unblocks the current work.

* Add libboost-program-options to the training docker image

This is needed by the `dedupe` program used in the `merge-corpus` step.

* Package toolchain binaries in tarballs

This avoids losing the execute bit on them when we download them in subsequent tasks.

* Minor bicleaner task fixes

- Upgrade CPU bicleaner tasks to `b-linux-large` (some datasets are very large - this lets them run in a reasonable amount of time)
- Fix artifact definitions. docker-worker doesn't like being given a relative path to the artifacts directory.

* Add `merge-corpus` pipeline step

The kind itself is quite boring -- all the interesting bits have been handled by transforms.

* Fix undeclared variables in find_upstreams.py

Co-authored-by: Gabriel Bustamante <gabrielbusta@gmail.com>

---------

Co-authored-by: Gabriel Bustamante <gabrielbusta@gmail.com>