Fix reverse locale detection in bicleaner script to work in Taskcluster
I believe this should work for both the existing pipeline and Taskcluster. The differenc in Taskcluster is that the bicleaner packs are downloaded from an earlier task, and don't contain the empty directory used currently to detect this condition. I did some spot checking, and these new check for the `.yaml` file seems like it should work.