Add `from_datasets:mono` transforms sequence
This new transforms sequence is very similar to the existing `per_dataset` one, except it's intended to be used with `mono-src` and/or `mono-trg` dataset types, and adds a new `locale` substitution based on which category is being used. (Note that we still want `src_locale` and `trg_locale` present for things like names though.)

This patch also fixes an oversight in `get_dataset_category` (now `get_dataset_categories`) that had it assuming that each dataset was only in one category (wrong!).