Add clean mono pipeline step (#127)
* Rename clean -> clean-corpus to prep for implementation of clean-mono

clean-mono & clean-corpus don't have enough in common that it makes sense to share a kind.

* Remove clean-mono parts from clean-corpus kind.

This patch is a big annoying to read - but it's mostly just renaming a few instances of `clean` -> `clean-corpus`, dropping the mono-specific fixes from `resources`, and moving everything into a single entry in `tasks` (because there's no need for per-provider behaviour at this point).

* Adjust dataset kind to publish news-crawl datasets with sanitized filenames.

In downstream kinds we've been using `dataset_sanitized` for filenames. We should do the same here, too.

* Add support for cache resources to be keyed by provider

This will be needed for `clean-mono`, as there's a number of provider-specific scripts that are run.

* Add `from_datasets:mono` transforms sequence

This new transforms sequence is very similar to the existing `per_dataset` one, except it's intended to be used with `mono-src` and/or `mono-trg` dataset types, and adds a new `locale` substitution based on which category is being used. (Note that we still want `src_locale` and `trg_locale` present for things like names though.)

This patch also fixes an oversight in `get_dataset_category` (now `get_dataset_categories`) that had it assuming that each dataset was only in one category (wrong!).

* Minor fixes for clean-mono.sh

Two things:
- Set `threads` to $(nproc) when `auto` is passed
- Explicitly run Python scripts with `python3` (needed for Taskcluster). This is already done for other jobs, so it ought to be fine for the existing pipeline...

* Add `clean-mono` kind

With the groundwork done in earlier commits, this is a pretty straightforward kind - just the usual substitutions, cache set-up, etc. being done.

* cache bust

* cache bust