Implement `finetune-teacher` step
Another step with ensembling/chunking. This step is almost a clone of `train-teacher` with the addition of consuming the model that comes of `train-teacher`. (In theory, this could've been put in the `train-teacher` kind, but I thought it best to separate it because we typically don't put parents/children in the same kind...)