Add support for training with `en` as the target language (#126)
* Add support for trg: en locale pairs

The only thing stopping this from working as far as the taskgraph is concerned was that we needed a bicleaner alias for the reverse locale pair.

I also ended up adjusting the prefix inside the archive. These packs come with a `$src-$trg` prefix, and it's much easier to script around them if we give them a consistent one.

* Set default locale pair to src: ru, trg: en

This patches the existing pipeline config for `test`.

* Fix reverse locale detection in bicleaner script to work in Taskcluster

I believe this should work for both the existing pipeline and Taskcluster. The differenc in Taskcluster is that the bicleaner packs are downloaded from an earlier task, and don't contain the empty directory used currently to detect this condition. I did some spot checking, and these new check for the `.yaml` file seems like it should work.

* Add `fr` to existing datasets where appropriate.

Mainly added to verify that `bicleaner-ai` works in reverse (it does).