Replace DAG.pdf with a more accurate version
While working on the Taskcluster pipeline I found some information missing from the DAG. Most notably, `merge_augmented` and `merge_translated` are not downstream of any training steps. This seems like it's happening because Snakemake can't calculate the DAG correctly when they are behind a `checkpoint`.To fix this, I temporarily hacked those checkpoints out with more hardcoded versions (link to the diff below).

I also noticed that the `evaluate` steps have an implicit dependency on `train_vocab` (it is referenced by the decoder ymls spit out by the `train` steps - but never explicitly noted in the Snakefile). I don't _think_ there's any downside to adjusting the Snakefile to account for this, so I've included it here.

The new graph is much, much uglier, of course - but it was also more reflective of reality.