Add `translate-mono-trg` pipeline step
The most notable part of this is that it's our first step that is parallelized many ways, and I've added a new transform to handle that.

Also of note is that while it seems like we should be adding `translate-mono-src` at the same time as this...we cannot. That step is actually downstream of this (despite earlier steps like `merge-mono` being a common ancestor of both).