Add toolchain build for browsermt-marian (#132)
This is a simple fork of the upstream marian project, so we can use the same build script for it. It looks like this was forked from an older version that we're using for upstream marian though, and I had to apply a few fixes to have it build properly - mostly around using a newer CUDA library.