Implement `collect-mono-trg` pipeline step
This step again breaks a bit of new ground: being the first step that requires unchunking (that is: pulling artifacts from many upstreams, and using all of them to generate its own artifact). This is handled by augmenting the `chunking` transform to allow the addition of N blocks to a task definition, with any necessary substitutions done in those blocks.