Implement `merge-augmented` pipeline step.
This step is the culmination of the `mono-trg` side of the DAG, and pulls in artifacts from a number of places to produce two artifacts that are used to power the teacher training part of the pipeline. There's no chunking, conditional dependencies, or even configuration from the training config to worry about - it's a very straightforward configuration.