Add split-{corpus,mono} pipeline steps (#134)
* Adjust `merge-mono` to build tasks for all locale pairs

My bad for missing this in review. Without it, we're unable to build `split-mono` tasks.

* Add Taskcluster support to split-{corpus,mono}.sh

In addition to the usual zstd support, this allows for optionally archiving all of the resulting files together. This is necessary in Taskcluster, because we must know the names of the files we need from other tasks prior to them running, and we don't know how many files this script will produce until runtime as things stand now.

* Add wrapper script for `split` pipeline step.

This is necessary because while the pipeline splits into files of N lines, while we run M chunks of `split` in Taskcluster. In cases where there are fewer files than chunks, we need to simulate success here, and publish an empty file (to ensure downstream steps don't break).

* Add split-{corpus,mono} pipeline steps