Use `provider_dataset` as `dataset` substitution.
I came to realize well after I originally wrote the early parts of the pipeline that `dataset` is usually in this form. It also turns out that this is important because certain steps (eg: cleaning) use the dataset name in this form to see if there's extra scripts to call out to for a given dataset.