Implement the remaining `mono-trg` parts of the pipeline in Taskcluster (#135)
* Add config.yml entry for workers with 4 GPUs

* Don't fail substitution if format values are missing

This allows us to easily format a string multiple times in different places.

* Add wrapper script for `translate` step.

Similar to the `split` steps, we need this wrapper to handle a case where a chunk has nothing to process.

* Add `translate-mono-trg` pipeline step

The most notable part of this is that it's our first step that is parallelized many ways, and I've added a new transform to handle that.

Also of note is that while it seems like we should be adding `translate-mono-src` at the same time as this...we cannot. That step is actually downstream of this (despite earlier steps like `merge-mono` being a common ancestor of both).

* Add zstd support to translate/{collect,merge-corpus}.sh

* Implement `collect-mono-trg` pipeline step

This step again breaks a bit of new ground: being the first step that requires unchunking (that is: pulling artifacts from many upstreams, and using all of them to generate its own artifact). This is handled by augmenting the `chunking` transform to allow the addition of N blocks to a task definition, with any necessary substitutions done in those blocks.

* Implement `merge-augmented` pipeline step.

This step is the culmination of the `mono-trg` side of the DAG, and pulls in artifacts from a number of places to produce two artifacts that are used to power the teacher training part of the pipeline. There's no chunking, conditional dependencies, or even configuration from the training config to worry about - it's a very straightforward configuration.