A collection of fixes for the dataset, clean, and merge stages of the pipeline (#136)
* zstd support and a minor fix for `download-mono.sh`

I originally bypassed this script in Taskcluster - but have since realized that it does some important filtering. Besides the zstd support being added, this commit ensures that `shuf` is fed an integer from `bc` (the version of `shuf` used in Taskcluster seems to treat it like an overflow if it's given a float).

* Download datasets through `download-{mono,corpus.sh}`

This isn't _strictly_ necessary for `corpus` since the wrapper doesn't do anything - but it may in the future, so we probably ought to use it.

* Make `command-context.from-parameters` optional

This allows us to use it in kinds where only some jobs are making use of the feature.

* Use `provider_dataset` as `dataset` substitution.

I came to realize well after I originally wrote the early parts of the pipeline that `dataset` is usually in this form. It also turns out that this is important because certain steps (eg: cleaning) use the dataset name in this form to see if there's extra scripts to call out to for a given dataset.

* Fix `merge-corpus` script to properly select input files

The original grep ended up matching substrings that weren't actually part of the locale....whoops.

* Add `bc` to `toolchain-docker` image

This is needed by the `download-mono.sh` script that we're now using in the `dataset` kind.

* Use command as part of cache key

This fixes a longstanding issue where changing the command a task runs doesn't invalidate caches. This is both wrong, and makes testing super annoying...

* Fix bicleaner threshold detection

Using `dataset_sanitized` doesn't work for datasets with a `/` in them. The new `dataset` substitution (that is the unprocessed provider and dataset) will always line up with what's in the config.