Get rid of `datasets` from `ci/config.yml`; simply `from_datasets` transform
When I wrote the early parts of the pipeline I thought it would be important that we explicitly record some basic data about datasets and locales in `ci/config.yml`. In the end, I don't think this is necessary or preferable for two reasons:
1) It makes training a new locale pair _very_ annoying. We typically have 50+ datasets, and the current data structure would mean we need a bunch of manual, error prone edits to this file (not to mention - the data structure would grow HUGE).
2) Adding all locale pairs to this config will end up with tens of thousands of tasks once we add just a handful of locale pairs.

The other reason this existed in the first place was to make sure we always generated some tasks without action parameters. Since the original work, we've added a good default set of parameters which will mean that `taskgraph tasks` will always generate a full set of tasks for `ru-en` - so there's no benefit to keeping them for this reason either.

Most of the changes to the kinds are very simple -- just a replacement of `from_datasets` with `task_substitution`. (In most kinds we were only using that transform to replace src/trg locale - which `task_substitution` can do just fine.)

The other notable change is big simplification of `from_datasets` -- ripping out most of its now unnecessary configuration, and pulling datasets from `parameters` instead.