Implement the remaining parts of the training pipeline in Taskcluster (#138)
* Add a substitution with shorter provider names

This is needed to make sure some Treeherder symbols stay under the 25 character limit.

* Implement `train-teacher` step

This is another chunked step. Its dependencies and arguments are different enough than the existing `train-backwards` step that I'm treating it as a separate kind.

* Implement `evaluate-teacher` step

As with `train-teacher`, this is chunked. Each of the clones of this consume its `train-teacher` counterpart with the same chunk number.

The chunking transforms had to be adjusted slightly to allow them to be used in kinds where only some jobs in the kind are chunked.

* Implement `finetune-teacher` step

Another step with ensembling/chunking. This step is almost a clone of `train-teacher` with the addition of consuming the model that comes of `train-teacher`. (In theory, this could've been put in the `train-teacher` kind, but I thought it best to separate it because we typically don't put parents/children in the same kind...)

* Implement `evaluate-finetuned-teacher` step

* Use separate wrapper scripts for `split-mono` and `split-corpus`

For two reasons:
1) They're different enough that it's a bit annoying to support both
2) More importantly - `split-corpus` is much further downstream in the graph than `split-mono` - and we don't want to bust caches for `split-mono` if something about the `split-corpus` part changes.

* Support simultaneous chunking and unchunking

Turns out we need to be able to do this! It's a great showcase for having multiple transform sequences in a single `transform`.

* Support `translate-corpus` in our pipeline wrapper script.

Unlike `split-{mono,corpus}` the differences here are nearly trivial -- it's unlikely this script will ever be changed in a way that only affects one of these kinds.

* Add `translate-corpus` pipeline step.

Notably, this kind is both chunked itself, and needs to unchunk and pull in artifacts from an upstream task that is chunked (`finetune-teacher`).

* Allow keys to have this_chunk/total_chunks substituted into them

This is necessary for `dependencies` definitions most notably.

* Generate less confusingly named files in `split-mono`

Naming these `src` is awkward, because they could be the `src` language or `trg` language. Easiest to just use a generic prefix.

* Implement `extract-best` pipeline step

* Rename collect kind to collect-mono

This was going to be used for `collect-corpus` originally...but these are going to be different enough that it doesn't makes sense to.

* Implement `collect-corpus` pipeline step

* Implement `translate-mono-src` pipeline step

* Implement `collect-mono-src` pipeline step

* Implement `merge-translated` pipeline step

* Support zstd in `score.sh`

* Implement `score` pipeline step

* Support zstd in `ce-filter.sh`

* Implement `cefilter` pipeline step

* Support zstd and automatic thread detection in alignments script

* Implement `alignments` pipeline step

* Support zstd in `train-student.sh`

* Implement `train-student` pipeline step

* Fix `fast-align` toolchain build to package tools in the root of the archive

* Update `train` and `toolchain` docker images with a few newly discovered necessities

* The `intel-mkl` tools - which are needed to build `browsermt-marian` correctly
* A couple of `libboost` libraries needed at runtime
* `pigz` - needed at the very end of the pipeline when we export the final artifacts

* Implement `evaluate-student` pipeline step

* Implement `finetune-student` pipeline step

* Implement `evaluate-finetuned-student` pipeline step

* Fix up `browsermt-toolchain` to be more usable

1) Use the same `toolchain-artifact`, so we can use the same build script
2) Include `extract_stats.py` in the archive if present (which it only is for `browsermt-marian`) - this is needed by one downstream step.

* Add requirements file for `quantize` pipeline step

This step relies on `numpy`.

* Implement `quantize` pipeline step

* Implement `evaluate-quantized` pipeline step

* Implement `export` pipeline step

* Implement `evalute` for `teacher-ensemble`

* Pass necessary marian args to `translate-mono-trg`

* Add comments about the two parts of the pipeline config that we don't yet support

* Bump max run time of most steps to 1 day

Many of these should probably be ratched down after we optimize instance types and see how long things tend to take in production -- but better to be too long than too short right now, to avoid running for hours and then killing something that's still valid.

* Get rid of `datasets` from `ci/config.yml`; simply `from_datasets` transform

When I wrote the early parts of the pipeline I thought it would be important that we explicitly record some basic data about datasets and locales in `ci/config.yml`. In the end, I don't think this is necessary or preferable for two reasons:
1) It makes training a new locale pair _very_ annoying. We typically have 50+ datasets, and the current data structure would mean we need a bunch of manual, error prone edits to this file (not to mention - the data structure would grow HUGE).
2) Adding all locale pairs to this config will end up with tens of thousands of tasks once we add just a handful of locale pairs.

The other reason this existed in the first place was to make sure we always generated some tasks without action parameters. Since the original work, we've added a good default set of parameters which will mean that `taskgraph tasks` will always generate a full set of tasks for `ru-en` - so there's no benefit to keeping them for this reason either.

Most of the changes to the kinds are very simple -- just a replacement of `from_datasets` with `task_substitution`. (In most kinds we were only using that transform to replace src/trg locale - which `task_substitution` can do just fine.)

The other notable change is big simplification of `from_datasets` -- ripping out most of its now unnecessary configuration, and pulling datasets from `parameters` instead.

* Implement dummy 'all' stage that ensures the entire pipeline will run

* Drop treeherder symbols for tasks that include dataset in their names because it is next to impossible to make them short enough to fit