Drop from_datasets:mono usage in kinds that don't run per dataset
This is a bug that was only uncovered when I did a run that had more than one dataset in the `mono-src` and `mono-trg` categories. Using this transform means we get one of these tasks _per_ dataset there (which is exactly the opposite of what we want -- these tasks are meant to operate on _all_ of the datasets in those categories).