Don't bust caches when non-training tasks are changed
This is accomplished by replacing the upstream `cached_tasks` transform with our own version (that re-uses as much as it can from upstream), that will avoid adding `docker-image`, `fetch`, and `toolchain` tasks to the cache digest.

We may want to take this even further - and never add _any_ dependencies to the digest...but I'm not prepared to make that leap until we have a bit more experience here. If we ever do that, we'll also need a way to override the cache (eg: something like `rebuild_builds` in the train action).