Follow-up fixes for Taskcluster pipeline after training a larger model (#139)
* Add new translations workers with larger disk space; different GPUs

* Bump cleaning & bicleaner to instances with more disk space; bicleaner-ai to one also with more GPUs

* Switch t-linux to b-linux for GPU worker types; bump translate-corpus to 1tb disk instances

Turns out translating the corpus can use massive amounts of space.

* Run dataset tasks on b-linux-large, to speed up compression/decompression

* Fix bicleaner 2.0

* Fix bicleaner stage selection

* Re-add {src,trg}_locale to many tasks to ensure stage targeting works

We lost this in a bunch of steps after the `from_datasets` transform stopped being used everywhere.

* Drop from_datasets:mono usage in kinds that don't run per dataset

This is a bug that was only uncovered when I did a run that had more than one dataset in the `mono-src` and `mono-trg` categories. Using this transform means we get one of these tasks _per_ dataset there (which is exactly the opposite of what we want -- these tasks are meant to operate on _all_ of the datasets in those categories).

* Pick up new version of taskgraph w/ group-by all & override for default task deadline

* Increase task deadline to 5 days

This needs to be much, much higher, but this is the maximum Taskcluster allows at this time.

* Don't bust caches when non-training tasks are changed

This is accomplished by replacing the upstream `cached_tasks` transform with our own version (that re-uses as much as it can from upstream), that will avoid adding `docker-image`, `fetch`, and `toolchain` tasks to the cache digest.

We may want to take this even further - and never add _any_ dependencies to the digest...but I'm not prepared to make that leap until we have a bit more experience here. If we ever do that, we'll also need a way to override the cache (eg: something like `rebuild_builds` in the train action).

* Add curl & wget to train image

We need wget for pre-downloading language identification models, and both are probably going to be useful at some point. May as well at them now, and avoid rebuilds later.

* Pre-download language identification model during cleaning to avoid potential race condition