Allow Taskcluster to run on pushes to any branch
Normally we try to restrict this to save on resources, but this repository builds very little on push (and what it does build is typically already cached). It sounds like we're going to use branches to help keep track of various experiments as well, so we'll need to be able to build from fairly arbitrary names. (If this becomes problematic at any point we can consider using a prefix for these experiment branches instead.)