Integrate new upstream chunking & from-deps transforms
Similar to the previous revision, most of this is fairly rote changes. Also similar is that `evaluate-finetuned-teacher` is moving to its own kind. This is happening in this case because `from-deps` does not allow multiple tasks in the same kind unless all are using `from-deps`.

The other things I should call out here are:
- Some usage of double curly braced strings in places where `task-context` must do the first substitution. I'm not super pleased about this, but I also couldn't find a workaround.
- The new `cast_to` transform is needed because the upstream `chunking` transform doesn't pull directly from parameters. `task-context` takes care of that, and converts ints to strings as it does so.