Update taskgraph; get rid of some custom transforms that are no longer necessary (#162)
* Upgrade to taskgraph 6.2.0

* Drop custom command-context & task-substitution transforms in favour of newly upstreamed `task-context`.

This is largely very boring, rote changes. The most notable part here is that `evaluate-quantized` is moving to its own kind. This is largely happening because I realized it was the only thing keeping `pipeline_args2` around - and moving it out made the rest of `evaluate` more readable.

* Integrate new upstream chunking & from-deps transforms

Similar to the previous revision, most of this is fairly rote changes. Also similar is that `evaluate-finetuned-teacher` is moving to its own kind. This is happening in this case because `from-deps` does not allow multiple tasks in the same kind unless all are using `from-deps`.

The other things I should call out here are:
- Some usage of double curly braced strings in places where `task-context` must do the first substitution. I'm not super pleased about this, but I also couldn't find a workaround.
- The new `cast_to` transform is needed because the upstream `chunking` transform doesn't pull directly from parameters. `task-context` takes care of that, and converts ints to strings as it does so.

* Fix a bug in taskcluster's split-corpus wrapper that caused src and ref chunks to sometimes be inconsistent when fewer files exist than there are chunks.

* Add comments to additional places where double curly braces are used