Add taskgraph-diff ci job (#276)
* Pass base_ref/base_rev to taskgraph

This is to make it possible to run `taskgraph --diff` accurately.

* Add Makefile target to generate taskgraph diffs

* Add taskgraph-diff ci job

These task diffs the full JSON graph of a couple of different training runs, comparing a PR against a base revision, or a push against the previous push.

It is an informational task designed not to fail even if there are differences. Any parameters that cause changes in the resulting taskgraph will cause an artifact with a diff of the before/after to be generated.