Use shorter version of `base_ref` for branches in the canonical repository
This fixes a regression caused by #276 that has caused decision tasks to fail for non `main` branches in this repository.

The "fix" is to refer to these branches by their short name (rather than `refs/heads/foo`). (This is not a great fix...but it's what we have to do to work around https://github.com/taskcluster/taskgraph/issues/380 for now.)