Make workers more easily configurable (fixes #254) (#277)
* Use more generic names for worker aliases

This is to make it easier to incorporate on-prem machines. Rather than have aliases that talk about GPU type, disk size, etc., we will be a bit abstract. This gives us the granularity we want in GCP, but also allows us to do things like just use "snakepit-workers" (or whatever) as the `worker-type` for _all_ GPU tasks when running on those machines.

* Move GPU & WORKSPACE definitions to config.yml

This moves all of the worker-specific configuration out of the kinds (yay).

The idea here is that we can maintain `worker-configuration` entries for each distinct type of machine, which the new transform will find and insert into the task definitions.