Switch CPU tasks to generic-worker/d2g images (fixes #473) (#533)
* Switch CPU tasks to generic-worker/d2g images (fixes #473)

This switches us from the deprecated docker-worker to generic-worker. generic-worker provides a translation layer for docker-worker tasks that avoids the need to change any payloads. (It will download specified images and run payload commands in them, rather than on the host machine.)

Upgrading to this new image will give us memory monitoring capabilities on the CPU workers because the new image has the GCP Ops Agent installed on it.

* Invalidate docker image cache to force rebuilds of all docker tasks on d2g workers