Re-enable generic-worker for CPU tasks
Analysis done in https://github.com/mozilla/firefox-translations-training/issues/538#issuecomment-2079685137 shows that the problems are intermittent, and largely related to spot terminations. We're seeing the latter in the existing workers anyways, so unless we find more serious issues with generic-worker for CPU tasks, we may as well go ahead with this.