merge-corpus: fetch bicleaner artifacts from `post-bicleaner-dummy` tasks
Each one of the `post-bicleaner-dummy` tasks can have artifacts from up to 98 bicleaner tasks. `merge-corpus` has room for up to 98 `post-bicleaner-dummy` tasks before it hits max dependencies, so we can handle ~9600 training datasets with this patch - enough to last awhile!

This is a slightly more exciting patch than I had hoped. When we pulled directly from `bicleaner` we could rely on simple attributes like `dataset_sanitized` on the upstream tasks to find the artifacts, because there was a 1:1 mapping between those attributes and the artifacts we want to fetch. Now that we're collected many bicleaner artifacts into a single `post-bicleaner-dummy` task, there's now a 1:M relationship, and we have to rely on a list of names in the attributes of those tasks. This necessitated writing a new transform to pull those out and add `fetches` entries for `merge-corpus`. It's nothing complicated, just annoying.