WIP switch GPU workers to image that uses multi engine generic worker
This will be necessary for #466 due to the "simple" engine not supporting chain of trust.

This initial work is just for some sanity checking that nothing obvious breaks. Aside from tasks in automation, we also need to see how interactive tasks behave, and if they meet the needs (most notably, see if we can have root in interactive tasks if needed).